+++ src/ModularPipelines/Modules/ModuleBase.cs
@@ -62,6 +62,7 @@ public abstract partial class ModuleBase
     
     internal Exception? Exception { get; set; }
 
+    internal abstract Task SetupModuleFromHistory(string? skipDecisionReason);
     internal abstract Task StartAsync();
     internal abstract void SetSkipped(SkipDecision skipDecision);
     internal abstract ModuleBase Initialize(IPipelineContext context);