+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,2 +1,3 @@
 *   Breaking: ShouldSkip virtual method now returns a `SkipDecision` class that allows you to plug in a reason
 *   Breaking: Refactor command logging with bit flag enums
+*   Try and setup modules from history if skipping due to categories